backend/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
greenlet==3.0.3
hanzidentifier==1.2.0
idna==3.7
jieba==0.42.1
numpy==2.0.1
pandas==2.2.2
pinyin-jyutping==0.9
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.32
sqlparse==0.5.1
tatoebatools==0.2.3
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
zhon==2.0.2
